[Server] Wrap creation of client with prices in a transaction
This operation involves creating a client and then creating and 
associating all prices with the newly generated clientId. So it is at 
the very least 2 querys.

The second query can fail if one of the prices is duplicated (same 
price, productId and clientId) after the first one succeded, in this 
case we want to rollback the creation of the client and all prices.

The user is still in the client creation page with all of its state 
intact so he can just correct whatever error and try again, the previous 
wrong client or prices won't conflict with this second try.